+++ apps/demo-app/services/generateImage.ts
@@ -1,4 +1,4 @@
-import { config, koldstart } from '@fal-ai/koldstart-client';
+import { config, koldstart } from '@fal/serverless-client';
 
 export type GenerateImageInput = {
   prompt: string;
@@ -19,7 +19,7 @@ export async function generateImage(
   input: GenerateImageInput
 ): Promise<ImageDataUri> {
   const result = await koldstart(
-    '552b4eab-7774-497d-926d-1d5b77c7c818/generate'
+    'github|38204337/a51c0ca0-9011-4ff0-8dc1-2ac0b42a9fd0/generate'
   ).run({
     input,
   });
