--- apps/demo-app/services/koldstart/credentials.ts
@@ -1,6 +1,0 @@
-import { Credentials } from '@fal-ai/koldstart-client';
-
-export const credentials: Credentials = {
-  id: '',
-  key: '',
-};
