+++ libs/client/package.json
@@ -1,7 +1,9 @@
 {
-  "name": "@fal-ai/koldstart-client",
+  "name": "@fal/serverless-client",
   "version": "0.0.1",
-  "files": [
-    "src/generated/server_pb.js"
-  ]
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/fal-ai/serverless-js",
+    "directory": "libs/client"
+  }
 }
