+++ tsconfig.base.json
@@ -15,8 +15,8 @@
     "skipDefaultLibCheck": true,
     "baseUrl": ".",
     "paths": {
-      "@fal-ai/koldstart-client": ["libs/client/src/index.ts"],
-      "@fal-ai/koldstart-codegen": ["libs/codegen/src/index.ts"]
+      "@fal/serverless-client": ["libs/client/src/index.ts"],
+      "@fal/serverless-codegen": ["libs/codegen/src/index.ts"]
     }
   },
   "exclude": ["node_modules", "tmp"]
